publish to discovery as endpoint -> uuid
Either an endpoint can change ip - unlikely in our env
or an endpoint can crash and restart with a new uuid.

having two uuids point to the same endpoint causes issues, if an
node restarts, the old uuid is not coming back, so replacing it
immediately is a good thing.

In the event that an IP does change, the connection to the old IP
will fail anyway, which wont harm anything.  This change may also be
required to support multi homed nodes.